Defer marking of objects as finalizable
This shifts responsibility for marking an object as "finalizable" from
object creation to object initialization.  We want to make the object
finalizable when Object.<init> completes.  For performance reasons we
skip the call to the Object constructor (which doesn't do anything)
and just take the opportunity to check the class flag.

Handling of clone()d object isn't quite right yet.

Also, fixed a minor glitch in stubdefs.

Bug 3342343

Change-Id: I5b7b819079e5862dc9cbd1830bb445a852dc63bf