[JIT] Fix profile trace dump
The debugging profile mode prints out a list of the top ten traces,
followed by recompilations.  In some cases, it is possible that a trace
was requested, but did not finish compiling before the run ended.  If
so, that could cause the dump to fail.  This CL adds a check for null
codeAddress to detect those cases.

Change-Id: I415fd94d8fa9e270f75d5114fa5cc5d993bd6997