New interpreter breakout mechanism
Introduce parallel handler entry points for mterp interpreters as a step
towards fully supporting debug, profile and JIT within mterp (instead of
bailing out to the portable debug interpreter).

This CL contains most of the structural changes that need to happen,
but does not yet enable the new switch mode.  In short, within the
mterp assembly interpreter register rIBASE points to an array of handlers
for Dalvik opcodes.  Instead of periodically checking for suspend,
debug, profiling and JIT trace selection breakouts, rIBASE may simply
be altered to point to the parallel breakout handlers when control needs
to be rerouted.  This will enable us to eliminate the separate portable
debug interpreter and the entire mechanism of switching between the fast
and portable interpreters.

The x86 implementation required a large number of changes because of the
need to dedicate a register to holding the table base.  It will now use %edx
(which was previously scratch).

Changes include:

  o Support for two styles of mterp assembly code generation: computed goto
    and jump table (ARM uses computed goto, x86 uses jump table)
  o New mterp config operators to trigger generation of alternate entry points.
  o Alternate entries route execution through new dvmCheckInst().  That's
    where the checking code will go.
  o For x86, reserved register edx as dedicated rIBASE.
  o For jump-table mterps, ignore "%break" operator and allow variable-sized
    handlers with no "sister" region.

Note that the x86-atom implementation will need substantial changes
to function in this new model.

Change-Id: I3a22048adb7dcfdeba4f94fbb977b26c3ab2fcb3