Clean up ref table dump
We were using the in-heap size as a sort key and a display item,
but the sizes of objects aren't all that interesting -- usually,
if the table is being dumped, it's because we ran into an object
count limitation rather than a total size limitation.

This largely eliminates the display of object sizes.  We show array
element counts for array objects.

Bug 3447435

Change-Id: I3bf31ed45317336ce9a41f058cce8eba88e58cc8