Pull the globals iteration into InitRefs.
This gives the new GC a way to iterate over the global object refs
defined in Globals.h, without having to have detailed knowledge of the
list of such refs. However, this change doesn't actually modify the GC
code in question, since it is not currently in a working state anyway.

InitRefs.c probably wants a slightly different name. Suggestions welcome.

Bonus: Cleaned up a vaguely-related clause in Class.c.

Change-Id: I406987c56ddf310c61d877574071c46f3717fc72