Fix some exception issues
The function that obtained an exception's message string was making a
bad assumption.  This has been corrected.

Also, in the invoke-object-init functions, we now test for a pending
exception after the call to dvmSetFinalizable().

Also, make the test for pending exception at the end of VM startup
an error rather than an assert.

Bug 4121213

Change-Id: I6912e5c79d63e8dda1a1dc2e788c7a8edcf487aa