Kill off code, and leave a note.
Based on discussion yesterday, the code in Copying.c was already
superfluous. I'm leaving dvmForEachRequiredReference(), except with a
note, just in case it ends up being needed. If by the time I'm done
with the next round of VM changes, it turns out not to be needed, then
I'll just go ahead and axe it.

Change-Id: Ib7cf06ba75851c7ffd98108c4c73cddee9259ddf