Interpreter fix - limit suspend checks
Register maps are generated only for a subset of instructions (to
limit memory consumption).  The previous interpreter restructuring
was doing suspend checks at locations that had no register maps.
This CL limits suspend checks to instructions that have register
maps, and also expands that set to include foward as well as
backwards branches.

Change-Id: Ia2b2e1096efe524fdb2a5fd4be171a4216b6a3b4