Normalize loop exit condition before checking counted loops
The loop exit condition for a simple one like "for (i=0; i<10; i++)"
can be done as
    - i<10
    - 10>i
    - !(i>=10)
    - !(10<=i)

This change normalizes the loop exit condition as if the exit path
is always on the taken direction of the branch. Doing so enables the
discovery of more counted loops without complicating the detecton logic.

Change-Id: I06f428334d4f1ca00d229fe76c282c6f161e8912