Fix method profiling [issue 4260697]
In the new interpreter control mechanism, method tracing invoke
events are generated at the call site, whereas debug entry events
happen prior to the interpretation of the first instruction of a
method.  Because we were failing to check for active profiling
on entry to a new interpreter instance, we would miss the profile
event for the first method interpreted.  We subsequently (and correctly)
recorded the return event, and ended up with a mismatch causing traceview
to complain and die.

Change-Id: Ibd93dfcd16c2f1fc62f6a86626a902eaf5d4b160