Remove DVM_LOCK_INIT and DVM_LOCK_INITIAL_THIN_VALUE.
The original implementation for thin locks used a magic non-zero value
to encode the initial thin lock state.  This magic value was kept
around in DVM_LOCK_INITIAL_THIN_VALUE and stored into the lock word of
newly allocated objects.  A later revision to the thin locking code
made the initial thin lock value be 0.  That change eliminated the
requirement that lock words be explicitly initialized as the allocator
always returns zero-filled memory.

Change-Id: I34e0b43b4c4db0f45cf7cf524e15d4a6096c1365