Remove an invalid assertion.
There are three situtations where insertions are performed into one of
the intern tables.  Two cases perform an insertion when an key value
pair is known to be absent.  One case performs an insert when a key
value pair might be present.  An assertion was added that errantly
checked that an insertion occured in the might be present case.  This
change leaves an assert in place for the absent cases and removes the
assert in the might be present case.

In addition, a comment has been improved to reinforce the condition of
the might be present insertion.

Change-Id: I84a9090a9ca338e164898e1d6893b2a23d74f5bc