Address CVE-2011-1090.
This adds a credentials check after connecting to the (alleged)
ADB control socket, to make sure that the other side really is
the system and not a malicious non-system process.

Since I was in the territory, I also did some minor comment
cleanup / correction (at the top of the file).

This change relies on a related change in system/core.

Change-Id: Iecbc4383cf285444f9d5a0414986951593777fa3