Eliminate the dvmHeapSourceGetPtrFlag interface.
This routine served answering questions about object addresses.  While
it was private to the garbage collector, its remaining use was inside
the heap profiler.  Within the heap profiler, a single path was
executed to tested a pointer for containment within the zygote.  In
its place, a new function, dvmIsZygoteObject, has been added to the GC
interface superseding the only use of dvmHeapSourceGetPtrFlag.

Change-Id: I347c9adb293bd6c196bbac3765015453133dc492