Fix PKIXCertPathValidatorSpiTest
Issue 17141

NOTE: This should not be merged into future releases.

Replace the contents of this test with those offered in the bug.
I thought about renaming the test to X500PrincipalTest, but there
is already another X500PrincipalTest. However, that test isn't
being pulled into CTS due to the build rules. I thought about
adding the tests under security.auth to pull it in, but that
would bring in some more tests that I don't know have been
reviewed. As a result, I've just modified the internals of this
test to match.

Change-Id: I9d7d1b8a202940588ee2475353917138526b1b42