Fix an issue regarding FindClass and threads without native methods.
Chapter 4 of the JNI specification states that when FindClass is called
through the Invocation Interface and there is no associated method, the
result of ClassLoader.getBaseClassLoader() is used as the class loader.

Previously, the case where FindClass is called from a main thread was
special cased to ensure this behavior.  However, threads which attach to
the VM but are not the main thread require similar treatment.  With this
change those threads are similarly treated as a special case.

Change-Id: Idbe33e02a10d248262b9e9f089b033ffe05c4706