Always grow the limit address of the a heap along with its size.
Previously, clearing the growth limit only updated the maximumSize of
the unlimited heap without updating the limit value.  This caused any
operations iterated over the extent of the heap to ignore allocations
beyond the original heap size.  With this change the limit is always
kept in agreement with the maximumSize.

Bug: 4539441
Change-Id: I08c8004e724549d196e1fd5d6ad22efc85b0c8d2